Project/requirements.txt
# Core Framework
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1

# Document Processing
pypdf2==3.0.1
python-docx==1.0.1
pdf2image==1.16.3
pytesseract==0.3.10
pillow==10.2.0

# RAG Components
langchain==0.3.0
langchain-community==0.0.16
sentence-transformers==2.5.0
chromadb==0.4.22

# LLM Integration
ollama==0.1.5

# Utilities
numpy==1.26.3
pandas==2.2.0
python-multipart==0.0.9

# Testing
pytest==7.4.3
httpx==0.25.2